Move ImageDirDataset to a shared location so that it can be used in multiple scripts.
